requirements.txt
rich
mysql-connector